+++ samplesheets/vue3app/src/utils/appUtils.ts
@@ -1,21 +1,27 @@
 // General app utilities
 
 import { useAppStore } from '@/stores/appStore.ts'
+import { REQ_GET } from '@/constants.ts'
 
 // Return standard Ajax request init parameters for fetch()
 export function getAjaxRequestInit (
-    method: string,
-    body: object
+    method?: string,
+    body?: object
 ): RequestInit {
   const appStore = useAppStore()
-  return {
-    method: method,
-    body: JSON.stringify(body),
-    credentials: 'same-origin',
-    headers: {
-      Accept: 'application/json',
-      'Content-Type': 'application/json',
-      'X-CSRFToken': appStore.sodarContext?.csrf_token || ''
-    }
+  if (!method) method = REQ_GET
+  const headers: HeadersInit = {
+    Accept: 'application/json',
+    'Content-Type': 'application/json',
   }
+  if (![REQ_GET, 'HEAD', 'OPTIONS'].includes(method.toUpperCase())) {
+    headers['X-CSRFToken'] = appStore.sodarContext?.csrf_token || ''
+  }
+  const ret: RequestInit = {
+    method: method,
+    credentials: 'same-origin',
+    headers: headers
+  }
+  if (body) ret.body = JSON.stringify(body)
+  return ret
 }